+++ cluster/clusterHandler.go
@@ -1,12 +1,15 @@
 package cluster
 
 import (
+	"encoding/json"
+	"io/ioutil"
 	"net/http"
 
 	gmux "github.com/gorilla/mux"
 	util "github.com/tmax-cloud/hypercloud-api-server/util"
 	caller "github.com/tmax-cloud/hypercloud-api-server/util/caller"
 	clusterDataFactory "github.com/tmax-cloud/hypercloud-api-server/util/dataFactory/cluster"
+	clusterv1alpha1 "github.com/tmax-cloud/hypercloud-multi-operator/apis/cluster/v1alpha1"
 	"k8s.io/klog"
 	// "encoding/json"
 )
@@ -62,7 +65,65 @@ func ListLNB(res http.ResponseWriter, req *http.Request) {
 	clmList, msg, status := caller.ListAccesibleCluster(userId, userGroups)
 	util.SetResponse(res, msg, clmList, status)
 	return
+}
 
+func InsertCLM(res http.ResponseWriter, req *http.Request) {
+	// queryParams := req.URL.Query()
+	vars := gmux.Vars(req)
+	namespace := vars["namespace"]
+	clustermanager := vars["clustermanager"]
+
+	if err := util.StringParameterException([]string{}, namespace, clustermanager); err != nil {
+		klog.Errorln(err)
+		util.SetResponse(res, err.Error(), nil, http.StatusBadRequest)
+		return
+	}
+	body, err := ioutil.ReadAll(req.Body)
+	if err != nil {
+		klog.Errorln(err)
+		util.SetResponse(res, err.Error(), nil, http.StatusBadRequest)
+		return
+	}
+
+	clm := &clusterv1alpha1.ClusterManager{}
+	if err := json.Unmarshal(body, clm); err != nil {
+		klog.Errorln(err)
+		util.SetResponse(res, err.Error(), nil, http.StatusBadRequest)
+	}
+
+	//INSERT_QUERY, item.Namespace, item.Cluster, item.MemberId, item.MemberName, item.Attribute, item.Role, item.Status, time.Now(), time.Now())
+	clmInfo := util.ClusterMemberInfo{
+		Namespace: clm.Namespace,
+		Cluster:   clm.Name,
+		MemberId:  clm.Annotations["owner"],
+		Attribute: "user",
+		Role:      "admin",
+		Status:    "owner",
+	}
+
+	// // 있는지 먼저 확인...
+	// var res string
+	// if res, err := clusterDataFactory.GetOwner(clmInfo); err != nil {
+	// 	msg := "Failed to get cluster owner from db"
+	// 	klog.Infoln(msg)
+	// 	util.SetResponse(res, msg, nil, http.StatusInternalServerError)
+	// 	return
+	// }
+	// // 있으니까 pass
+	// if res != "" {
+
+	// }
+
+	if err := clusterDataFactory.Insert(clmInfo); err != nil {
+		msg := "Failed to delete cluster info from db"
+		klog.Infoln(msg)
+		util.SetResponse(res, msg, nil, http.StatusInternalServerError)
+		return
+	}
+	msg := "Success to delete cluster info from db"
+	klog.Infoln(msg)
+	util.SetResponse(res, msg, nil, http.StatusOK)
+	return
 }
 
 func DeleteCLM(res http.ResponseWriter, req *http.Request) {