+++ push.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-img=192.168.9.12:5000/hypercloud-api-server:v0.24
+img=192.168.9.12:5000/hypercloud-api-server:v3.10
 docker rmi $img 
 docker build -t $img  . 
 docker push $img